Rich text: move delete handling to ref callback
